gumbo-rest-service/pyproject.toml
[tool.poetry]
name = "gumbo-rest-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "gumbo_rest_service"}]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.9"
pandas = "^1.1.0"
fastapi = "^0.109.0"
python-dotenv = "^1.0.0"
gumbo-dao = {path = "../gumbo-dao"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
